requirements.txt
PySide6>=6.5.0
PyMuPDF>=1.23.0
pypdf>=3.0.0
reportlab>=4.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
pikepdf>=8.0.0
pytest>=7.0.0
